added `javascript` utility for script tags and introduced a `runtime` chunk
The code still needs cleanup. And we also need to think hard about this setup. The current version introduces a runtime chunk as soon as we have common chunks. This can probably be improved upon a bit, but that is something we need to think about. It introduces an extra file but the alternative is very complex (if it is even possible) when solved generically. We could simplify that complex situation to only introduce a `runtime` chunk when there is more than one common.